fix test_docker_storage_mounts
